fix: refactor container inverter
remove alot of dead code, improve docs, uipdate types
