+++ packages/nextjs-mf/src/plugins/container/InvertedContainerPlugin.ts
@@ -1,0 +1,71 @@
+import type { Chunk, Compiler } from "webpack";
+//@ts-ignore
+import type { ModuleFederationPluginOptions } from "../types";
+import InvertedContainerRuntimeModule from "./InvertedContainerRuntimeModule";
+
+/**
+ * Interface for InvertedContainerOptions, extending ModuleFederationPluginOptions.
+ * This interface includes additional fields specific to the plugin's behavior.
+ */
+interface InvertedContainerOptions extends ModuleFederationPluginOptions {
+  remotes: Record<string, string>; // A map of remote modules to their URLs.
+  name?: string; // The name of the current module.
+  verbose?: boolean; // A flag to enable verbose logging.
+}
+
+/**
+ * InvertedContainerPlugin is a Webpack plugin that handles loading of chunks in a federated module.
+ * It sets up runtime modules for each chunk, ensuring the proper loading of remote modules.
+ */
+class InvertedContainerPlugin {
+  private options: InvertedContainerOptions;
+
+  /**
+   * Constructor for the InvertedContainerPlugin.
+   * @param {InvertedContainerOptions} options - Plugin configuration options.
+   */
+  constructor(options: InvertedContainerOptions) {
+    this.options = options || ({} as InvertedContainerOptions);
+  }
+
+  /**
+   * Apply method for the Webpack plugin, handling the plugin logic and hooks.
+   * @param {Compiler} compiler - Webpack compiler instance.
+   */
+  apply(compiler: Compiler) {
+    // Hook into the compilation process.
+    compiler.hooks.thisCompilation.tap(
+      'InvertedContainerPlugin',
+      (compilation) => {
+        // Create a WeakSet to store chunks that have already been processed.
+        const onceForChunkSet = new WeakSet();
+
+        // Define a handler function to be called for each chunk in the compilation.
+        const handler = (chunk: Chunk, set: Set<string>) => {
+          // If the chunk has already been processed, skip it.
+          if (onceForChunkSet.has(chunk)) return;
+
+          // Mark the chunk as processed by adding it to the WeakSet.
+          onceForChunkSet.add(chunk);
+
+          // Add an InvertedContainerRuntimeModule to the chunk, which handles
+          // the runtime logic for loading remote modules.
+          compilation.addRuntimeModule(
+            chunk,
+            new InvertedContainerRuntimeModule(set, this.options, {
+              webpack: compiler.webpack,
+            })
+          );
+        };
+
+        // Hook the handler function into the compilation process.
+        compilation.hooks.additionalChunkRuntimeRequirements.tap(
+          'InvertedContainerPlugin',
+          handler
+        );
+      }
+    );
+  }
+}
+
+export default InvertedContainerPlugin;
