+++ packages/nextjs-mf/src/plugins/container/types.ts
@@ -1,7 +1,5 @@
-import type { container } from 'webpack';
+import type { container } from "webpack";
 
 export type ModuleFederationPluginOptions = ConstructorParameters<
   typeof container.ModuleFederationPlugin
 >['0'];
-
-export type RemotesObject = ModuleFederationPluginOptions['remotes'];
